feature: highlight current train when viewing station arrivals
When you're on a train, and you click into the train view, you can see
when your train will arrive at a given station.

If you need to transfer, you often will want to view that station
to see when the train you need to catch will arrive.

This makes that workflow a bit easier with two changes:

1. makes it possible to tap on a station name in the "train arrivals"
   view to go to a station

2. when you do that, pass the current train ID to the next view,
   allowing the UI to highlight your train in the list of arrivals.

This PR is missing some styles which I leave up to the benevolent
marta.io dictator to decide:

1. do the station names in the train arrival time list need to
   indicate they are links somehow?

2. what should the UI indicator be for 'this is the train you are on'
   in the station view?